requirements.txt
Django>=3.0.5,<3.1.0
djangorestframework>=3.11.0,<3.12.0
djangorestframework-simplejwt>=4.4.0,<4.5.0
psycopg2>=2.8.5,<2.9.0
redis>=3.4.1,<3.5.0
ldap3>=2.7,<2.8

flake8>=3.7.9,<3.8.0